+++ packages/client/src/shared/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useToggle } from './useToggle'
+export { useForm } from './useForm'
