+++ packages/client/src/widgets/SignUpForm/ui.tsx
@@ -2,7 +2,7 @@ import { Form, FormInput } from '@/shared/ui/Form'
 import { Button } from '@/shared/ui/Button'
 import { useNavigate } from 'react-router-dom'
 import authServices from '@/shared/services/authServices'
-import useForm from '@/shared/hooks/useForm'
+import { useForm } from '@/shared/hooks'
 import schema from './schema'
 import type { TUseForm } from './types'
